Fix __hash__ return type and tighten composition tolerance
- Change PyElement.__hash__ return type from u64 to isize (PyO3 expects signed)
- Tighten compositions_almost_equal relative tolerance from 10% to 1%
- Add singular matrix test for get_lattice_inv_matrix

+++ extensions/rust/src/python.rs
@@ -811,8 +811,8 @@ impl PyElement {
         self.inner == other.inner
     }
 
-    fn __hash__(&self) -> u64 {
-        self.inner.atomic_number() as u64
+    fn __hash__(&self) -> isize {
+        self.inner.atomic_number() as isize
     }
 }
